requirements.txt
Django==2.0.5
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
django-ckeditor==5.4.0
